backend/requirements.txt
PyMySQL~=1.0.2
DBUtils~=1.2
pandas~=1.2.0
exobject~=0.3
tqdm~=4.55.0
numpy~=1.20.0
pyarrow~=3.0.0
scipy~=1.7.0
elasticsearch~=7.8.1
func-timeout~=4.3.5
requests~=2.25.1
SQLAlchemy~=1.4.35
APScheduler~=3.7.0
aioredis~=2.0.1
sqlacodegen~=2.3.0
fastapi~=0.78.0
python-multipart
pyjwt
uvicorn
redis~=3.5.3
zmq
betterproto
uvicorn[standard]
protobuf~=4.22.1
fsspec==2022.7.1
kafka-python==2.0.2